// src/types/pagesTypes.ts
export type AboutProps = {
  darkMode: boolean;
  approachTwo: Array<object>;
  teamMembers: Array<object>;
  clientsData: Array<object>;
};
export type ContactProps = {
  darkMode: boolean;
  locationData: Array<object>;
};
export type HomeProps = {
  darkMode: boolean;
  approachData: Array<object>;
  clientsData: Array<object>;
  innovationData: Array<object>;
  productsData: Array<object>;
};
export type NotFoundProps = {
  darkMode: boolean;
};
export type ServiceProps = {
  darkMode: boolean;
  household: Array<object>;
};